# Conflicts:
#	src/main/java/org/example/controller/StudentController.java
#	src/main/resources/data.sql

+++ src/main/java/org/example/dto/CheckSemGradeDto.java
@@ -13,7 +13,7 @@ public class CheckSemGradeDto {
     private String semYear;
     private String semSem;
     private Long semCrecpl;
-    private Long semAvgCrecpl;
+    private Float semAvgCrecpl;
 //    private String grade;
 //    private Long lectCredit;
